+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ## Unreleased
 
+### Added
+- Documented ``Stop.NONE`` and ``Stop.COAST_SMART``.
+
 ### Changed
 - Changed `PrimeHub.display.image()` to `PrimeHub.display.icon()` and renamed
   its kwarg from `image` to `icon`.
+- Improved presentation and docstrings of the ``ubuiltins`` module.
 
 ## 3.2.0b5 - 2022-11-11